refactor(cli): extract git + repo-id helpers — phase 3 redo (card cb8d8990)
Same extraction pattern as phases 1 (#1063) and 2 (#1064): hard-reset to current rust-rewrite tip, re-extract spawn_claim_worktree + git_show_format + git_rev_parse_branch + cwd_github_repo_id + parse_github_repo_id + slugify into work_commands_git.rs. Cross-module callees in work_commands_gh.rs's open_pr_and_link routed through crate::work_commands_git::.

work_commands.rs: 1713 → 1521 LOC (-192)
work_commands_git.rs: 0 → ~205 LOC

20 tests pass; clippy clean (with the pre-existing #1058-pending pr_create_base_branch allow); fmt clean.

Original #1047 closed due to post-#1049 cross-module call conflict.
