+++ crates/airc-cli/src/main.rs
@@ -79,6 +79,7 @@ mod update_commands;
 mod work_cli;
 mod work_commands;
 mod work_commands_gh;
+mod work_commands_git;
 mod work_suggestions;
 mod workspace_cli;
 mod workspace_commands;
